Deprecate RadioButton component
